+++ wscheck/checker.py
@@ -1,12 +1,12 @@
 import re
 
 RULES = {
-    'WSW001': 'Bad line ending',
-    'WSW002': 'Tailing whitespace',
-    'WSW003': 'Indentation is not multiple of 2',
-    'WSW004': 'Indentation with non-space character',
-    'WSW005': 'No newline at end of file',
-    'WSW006': 'Too many newline at end of file',
+    'WSC001': 'Bad line ending',
+    'WSC002': 'Tailing whitespace',
+    'WSC003': 'Indentation is not multiple of 2',
+    'WSC004': 'Indentation with non-space character',
+    'WSC005': 'No newline at end of file',
+    'WSC006': 'Too many newline at end of file',
 }
 
 
@@ -75,15 +75,15 @@ class WhitespaceChecker(object):
         for line, line_text_eol in enumerate(lines, start=1):
             line_text, line_eol = line_text_eol
 
-            if 'WSW001' in self._rules:
+            if 'WSC001' in self._rules:
                 if not line_eol == '' and not line_eol == '\n':
-                    self._add_issue(rule='WSW001', path=file_path, line=line, col=len(line_text) + 1, context=line_text,
+                    self._add_issue(rule='WSC001', path=file_path, line=line, col=len(line_text) + 1, context=line_text,
                                     message_suffix='{!r}'.format(line_eol))
 
-            if 'WSW002' in self._rules:
+            if 'WSC002' in self._rules:
                 tailing_whitespace_match = self._TAILING_WHITESPACE_TEMPLATE.search(line_text)
                 if tailing_whitespace_match is not None:
-                    self._add_issue(rule='WSW002', path=file_path, line=line, col=tailing_whitespace_match.start() + 1, context=line_text)
+                    self._add_issue(rule='WSC002', path=file_path, line=line, col=tailing_whitespace_match.start() + 1, context=line_text)
 
             if line_text.strip() == '':
                 continue
@@ -92,14 +92,14 @@ class WhitespaceChecker(object):
             if indent_match is not None:
                 line_indent = indent_match.group()
 
-                if 'WSW003' in self._rules:
+                if 'WSC003' in self._rules:
                     if not len(line_indent.replace('\t', '    ')) % 2 == 0:
-                        self._add_issue(rule='WSW003', path=file_path, line=line, col=len(line_indent) + 1, context=line_text)
+                        self._add_issue(rule='WSC003', path=file_path, line=line, col=len(line_indent) + 1, context=line_text)
 
-                if 'WSW004' in self._rules:
+                if 'WSC004' in self._rules:
                     character_match = self._NOT_SPACES_TEMPLATE.search(line_indent)
                     if character_match is not None:
-                        self._add_issue(rule='WSW004', path=file_path, line=line, col=character_match.start() + 1,
+                        self._add_issue(rule='WSC004', path=file_path, line=line, col=character_match.start() + 1,
                                         context=line_text)
 
     def _add_issue(self, rule, path, line, col, context, message_suffix=None):
@@ -128,14 +128,14 @@ class WhitespaceChecker(object):
                 break
             empty_lines += 1
 
-        if 'WSW005' in self._rules:
+        if 'WSC005' in self._rules:
             if empty_lines == 0:
                 line_text = lines[-1][0]
-                self._add_issue(rule='WSW005', path=file_path, line=len(lines), col=len(line_text) + 1, context=line_text)
+                self._add_issue(rule='WSC005', path=file_path, line=len(lines), col=len(line_text) + 1, context=line_text)
 
-        if 'WSW006' in self._rules:
+        if 'WSC006' in self._rules:
             if empty_lines > 1:
                 shift = min(len(lines), empty_lines + 1)
                 line_text = lines[-shift][0]
-                self._add_issue(rule='WSW006', path=file_path, line=len(lines) - shift + 1, col=len(line_text) + 1,
+                self._add_issue(rule='WSC006', path=file_path, line=len(lines) - shift + 1, col=len(line_text) + 1,
                                 context=line_text, message_suffix='(+{})'.format(empty_lines - 1))